refactor(data): NearbySearchの関数定義を`nearbysearch.go`に移動
