+++ internal/infrastructure/api/google/places/nearbysearch.go
@@ -4,10 +4,59 @@ import (
 	"context"
 	"fmt"
 	"go.uber.org/zap"
+	"poroto.app/poroto/planner/internal/domain/utils"
 	"time"
 
 	"googlemaps.github.io/maps"
 )
+
+type FindPlacesFromLocationRequest struct {
+	Location    Location
+	Radius      uint
+	Language    string
+	Type        *maps.PlaceType
+	SearchCount int
+}
+
+func (r PlacesApi) FindPlacesFromLocation(ctx context.Context, req *FindPlacesFromLocationRequest) ([]Place, error) {
+	var placeType maps.PlaceType
+	if req.Type != nil {
+		placeType = *req.Type
+	}
+
+	placeSearchResults, err := r.nearBySearch(ctx, &maps.NearbySearchRequest{
+		Location: &maps.LatLng{
+			Lat: req.Location.Latitude,
+			Lng: req.Location.Longitude,
+		},
+		Radius:   req.Radius,
+		Language: req.Language,
+		Type:     placeType,
+	}, req.SearchCount)
+	if err != nil {
+		return nil, err
+	}
+
+	// Getting places nearby
+	var places []Place
+	for _, place := range placeSearchResults {
+		places = append(places, createPlace(
+			place.PlaceID,
+			place.Name,
+			place.Types,
+			place.Geometry,
+			place.Photos,
+			place.OpeningHours != nil && place.OpeningHours.OpenNow != nil && *place.OpeningHours.OpenNow,
+			place.Rating,
+			place.UserRatingsTotal,
+			utils.StrOmitEmpty(place.FormattedAddress),
+			utils.StrOmitEmpty(place.Vicinity),
+			place.PriceLevel,
+		))
+	}
+
+	return places, nil
+}
 
 // nearBySearch Places API Nearby Search
 // https://developers.google.com/maps/documentation/places/web-service/search-nearby
