+++ internal/infrastructure/api/google/places/places.go
@@ -1,7 +1,6 @@
 package places
 
 import (
-	"context"
 	"fmt"
 	"go.uber.org/zap"
 	"os"
@@ -40,52 +39,4 @@ func NewPlacesApi() (*PlacesApi, error) {
 		mapsClient: c,
 		logger:     logger,
 	}, nil
-}
-
-type FindPlacesFromLocationRequest struct {
-	Location    Location
-	Radius      uint
-	Language    string
-	Type        *maps.PlaceType
-	SearchCount int
-}
-
-func (r PlacesApi) FindPlacesFromLocation(ctx context.Context, req *FindPlacesFromLocationRequest) ([]Place, error) {
-	var placeType maps.PlaceType
-	if req.Type != nil {
-		placeType = *req.Type
-	}
-
-	placeSearchResults, err := r.nearBySearch(ctx, &maps.NearbySearchRequest{
-		Location: &maps.LatLng{
-			Lat: req.Location.Latitude,
-			Lng: req.Location.Longitude,
-		},
-		Radius:   req.Radius,
-		Language: req.Language,
-		Type:     placeType,
-	}, req.SearchCount)
-	if err != nil {
-		return nil, err
-	}
-
-	// Getting places nearby
-	var places []Place
-	for _, place := range placeSearchResults {
-		places = append(places, createPlace(
-			place.PlaceID,
-			place.Name,
-			place.Types,
-			place.Geometry,
-			place.Photos,
-			place.OpeningHours != nil && place.OpeningHours.OpenNow != nil && *place.OpeningHours.OpenNow,
-			place.Rating,
-			place.UserRatingsTotal,
-			utils.StrOmitEmpty(place.FormattedAddress),
-			utils.StrOmitEmpty(place.Vicinity),
-			place.PriceLevel,
-		))
-	}
-
-	return places, nil
 }
